% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{base_53_hierarchy}
\alias{base_53_hierarchy}
\title{Assign broad immune lineage labels for 53-state reference names}
\usage{
base_53_hierarchy(state.labels)
}
\arguments{
\item{state.labels}{Character vector of state names.}
}
\value{
data.frame with state_name, subtype, major_lineage.
}
\description{
Assign broad immune lineage labels for 53-state reference names
}
