% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_experiment1.R
\name{extract_sample_truth_from_names}
\alias{extract_sample_truth_from_names}
\title{Extract simulation truth from sample names}
\usage{
extract_sample_truth_from_names(
  sample.names,
  truth.seq = seq(0.005, 0.2, by = 0.005)
)
}
\arguments{
\item{sample.names}{Character vector of sample names.}

\item{truth.seq}{Numeric vector of target proportions.}
}
\value{
data.frame with Sample, TargetCell, Index and Truth.
}
\description{
Assumes sample names such as CD4Tn_1, CD4Tn_2, ..., where the numeric suffix
indexes target proportions seq(0.005, 0.2, by = 0.005).
}
