% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{ImmuneHierDeconv}
\alias{ImmuneHierDeconv-package}
\alias{ImmuneHierDeconv}
\title{ImmuneHierDeconv: ImmuCellAI 2.0 immune deconvolution}
\description{
Deterministic variational Bayesian deconvolution of bulk RNA-seq profiles
into 53 relative immune-cell fractions, with targeted hierarchical
refinement of CD4 and CD8 T-cell states.
}
\author{
\strong{Maintainer}: Jingmin Yang \email{yangjingmin89@gmail.com}

Authors:
\itemize{
  \item Jingmin Yang \email{yangjingmin89@gmail.com}
}

}
