% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_experiment1.R
\name{map_simulation_target_to_states}
\alias{map_simulation_target_to_states}
\title{Map simulation target names to reference state names}
\usage{
map_simulation_target_to_states(target)
}
\arguments{
\item{target}{Character vector of target cell names from sample names.}
}
\value{
Named list when target has length > 1, otherwise a character vector.
}
\description{
Map simulation target names to reference state names
}
