% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_experiment1.R
\name{map_target_to_bayesprism_panel}
\alias{map_target_to_bayesprism_panel}
\title{Map target state names to BayesPrism-panel names}
\usage{
map_target_to_bayesprism_panel(x)
}
\arguments{
\item{x}{Character vector of target cell names.}
}
\value{
Character vector of mapped target names.
}
\description{
Map target state names to BayesPrism-panel names
}
