% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_reference.R
\name{merge_reference}
\alias{merge_reference}
\title{Merge reference profiles}
\usage{
merge_reference(reference, merge.groups)
}
\arguments{
\item{reference}{Gene x state matrix.}

\item{merge.groups}{Named list. Names are new merged states; values are original states.}
}
\value{
Merged reference matrix.
}
\description{
Merge reference profiles
}
